improves integration tests to inject context headers
